Add CRUD

+++ src/main/java/services/RoleService.java
@@ -2,10 +2,7 @@ package services;
 
 import models.Role;
 import org.hibernate.Session;
-import org.hibernate.Transaction;
 import org.hibernate.query.Query;
-import utils.HibernateSessionUtils;
-import viewmodels.RoleCreateVM;
 
 import java.util.List;
 
@@ -16,47 +13,58 @@ public class RoleService {
         this.context = context;
     }
 
-    public Role GetById(int id){
+    public Role GetById(int id) {
         Role role = context.get(Role.class, id);
         return role;
     }
 
-    public Role GetByName(String name){
-        Role role = context.load(Role.class, name);
-        return role;
+    public void CreateRole(Role model) {
+        try {
+            Role role = new Role();
+            role.setName(model.getName());
+            context.save(role);
+            context.beginTransaction().commit();
+        } catch (Exception ex) {
+            System.out.println("Create role error: " + ex.getMessage());
+        }
     }
 
-    public void CreateRole(RoleCreateVM model){
-        Role role = new Role();
-        role.setName(model.getName());
-        context.save(role);
-    }
-
-    public List<Role> GetAllRoles(){
-        Query query = context.createQuery("FROM Role");
-        List<Role> list = query.list();
-        return list;
+    public List<Role> GetAllRoles() {
+        try {
+            Query query = context.createQuery("FROM Role");
+            List<Role> list = query.list();
+            return list;
+        } catch (Exception ex) {
+            System.out.println("Get roles error: " + ex.getMessage());
+            return null;
+        }
     }
 
     public void UpdateRole(Role model) {
-        context.update(model);
+        try {
+            context.update(model);
+            context.beginTransaction().commit();
+        } catch (Exception ex) {
+            System.out.println("Update role error: " + ex.getMessage());
+        }
+
     }
 
     public void DeleteRole(Role model) {
-        Transaction tx = context.beginTransaction();
-        context.remove(model);
-        tx.commit();
+        try {
+            context.delete(model);
+            context.beginTransaction().commit();
+        } catch (Exception ex) {
+            System.out.println("Delete role error: " + ex.getMessage());
+        }
     }
 
     public void DeleteRole(int id) {
         Role role = this.GetById(id);
-        this.DeleteRole(role);
-    }
-
-    public void DeleteRole(String name) {
-        Role role = this.GetByName(name);
-        System.out.print("Role: ");
-        System.out.println(role);
-        this.DeleteRole(role);
+        if (role != null) {
+            this.DeleteRole(role);
+        } else {
+            System.out.println("Role not found");
+        }
     }
 }
